Move resource scanning into it's own module
